+++ CHANGELOG.md
@@ -1,8 +1,21 @@
 ## master
 
-## 1.2.8 (April 2, 2017)
+
+## 1.2.12 (March 31, 2017)
 
 * Bug fix.
+
+## 1.2.11 (March 31, 2017)
+
+* Using current package.json and updating info
+
+## 1.2.10 (March 30, 2017)
+
+* Add npm badget to README.md
+
+## 1.2.9 (March 15, 2017)
+
+* Add CHANGELOG.md template.
 
 ## 1.0.0 (March 15, 2017)